Add early stopping to Primal Dual algorithm
Add the ability to stop the Primal Dual (PD) algorithm after a fixed
number of iterations, after which the solver switches to using the
Successive Shortest Paths (SSP) algorithm. This is often useful in
practice because the PD algorithm may exhibit a "long tail" behavior
where it spends an inordinate amount of time resolving the last few
excess nodes, whereas the SSP algorithm can handle a small number of
excess nodes relatively efficiently.

It might be preferable to implement a more intelligent early-stopping
condition, such as checking when the number of excess nodes addressed
per iteration falls below some threshold. Will save that for a future
enhancement, though.